requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
idna==2.10
lxml==4.6.2
pyaml==20.4.0
pymongo==3.11.2
PyYAML==5.4
requests==2.25.1
soupsieve==2.1
urllib3==1.26.2
